backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
python-dotenv==1.0.0
pydantic==2.9.0
pydantic-settings==2.5.0
google-genai==0.2.2
librosa==0.10.1
soundfile==0.12.1
numpy==1.26.4
python-multipart==0.0.17

